+++ CHANGES.md
@@ -15,6 +15,8 @@ Unreleased
 - Use modern `pyproject.toml` project metadata. Use flit_core instead of setuptools as
   build backend.
 - Use `datetime.now(timezone.utc)` instead of deprecated `datetime.utcnow`. #758
+- Never look at the `X-Forwarded-For` header, always use `request.remote_addr`,
+  requiring the developer to configure `ProxyFix` appropriately. #700
 
 
 Version 0.6.3